Add ability to specify expected vector set for grading.
This allows you to specify 'expected_vectors' setting when setting in vectordraw_settings.

The entry should be a dictionary of vector names to vector properties that should be checked.
The supported properties are tail, tip, length, and angle.
Each property accepts a tolerance setting - tail_tolerance, tip_tolerance,
length_tolerance, angle_tolerance. If the tolerance is not given, it defaults
to a predefined tolerance for the given property.

Example:

    expected_vectors: {
      'N': {length: 5, tail: [0, 0], tail_tolerance: 0.5},
      'f': {angle: 180, tail: [0, 0], angle_tolerance: 5}
    }