Rewrite the check_angle grading function.
It makes more sense conceptually to calculate the angle
between the given vector and the unit vector with the expected angle
and check that the size of the calculated angle is within the accepted tolerance.

This also fixes a bug where check_angle wouldn't work correctly
when the expected angle was negative.