Make segment tails draggable.
Vectors often start at the same origin point, which is why vector tails
are not draggable - when you press on a tail, you start drawing a new vector
rather than dragging the existing one, because drawing a set of vectors
all starting at the same origin point would be too annoying otherwise.

Segments usually don't start at the same point, so it makes sense to allow
draggin the tails (which for segments are intistinguishable from tips from
the users' perspective).